Move redux utilities to dedicated directory
