+++ src/core/index.ts
@@ -1,1 +1,1 @@
-export { store } from './store';
+export { store } from './services/redux/store';
