+++ src/core/services/redux/user/user.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import type { RootState } from '../../store';
+import type { RootState } from '../store';
 
 interface User {
   email: string;
